Simplify validation and make it more verbose
Checking if a number is possible didn't take the number type into account anyway, since in
`test_number_length/2` we were always passing `PhonenUmberTypes.unknown()`. This change prevents
the code from pretending like it does support validating by types. If we want to re-introduce it,
we should to so properly.